Provide a URL/ServerUrl extension, and extension loading
This patch was prompted by a desire to have `url` and `serverurl` functions that
proxy to the zend-expressive-helpers classes. In order to do that, however, I
ended up needing the following as well:

- The ability to register the `League\Plates\Engine` as a service, and consume
  it when present.
- A factory for `League\Plates\Engine` that can load extensions into the engine
  (via instance, service name, or class name), as well as load the custom URL
  extension by default.
- The UrlExtension itself, which composes UrlHelper and ServerUrlHelper
  instances.
- A factory for the UrlExtension that raises exceptions if either helper is
  missing.

I updated the PlatesRendererFactory to delegate to the new PlatesEngineFactory
if no `League\Plates\Engine` service is present in order to create the Plates
engine instance. This also then ensures that by default, we have the
`UrlExtension` loaded.