Ensure we raise an InvalidExtensionException for invalid extension types
Adds a test for the fix originally provided in #9, and an additional
test for cases where a non-service class name was provided; each should
raise the exception, in order to prevent a `TypeError` (PHP 7) or fatal
error (PHP 5) from occurring instead.

Revised the logic to prevent duplication of the check against
`ExtensionInterface`.